pyproject.toml
[project]
name = "cloud-stuff"
version = "0.1.0"
description = ""
readme = "README.md"
package-mode = false
requires-python = ">=3.9"

dependencies = [
  "requests >= 2.31.0",
  "supabase >= 2.4.1",
  "python-dotenv >= 1.0.1",
  "sqlalchemy >= 2.0.29",
  "urllib3 <= 2.0.0",
  "psycopg2-binary >= 2.9.9",
  "sqlalchemy-libsql >= 0.1.0",
]

[dependency-groups]
dev = [
  "pre-commit >= 4.2.0",
  "pytest >= 8.1.1",
  "pytest-cov >= 5.0.0",
]
